Added a scaling variable to the staticmodel entity class.
Put in the final ground (although Robert needs to flip some normals) which is scaled 50%.
The ground's physics vertices also use this scaling.

Put in a check in the AnimatedModel and StaticModel effects so if a vertex normal is facing away from the light, it doesn't get calculated by the shadow map effect.